MCR-3080 rename MCRObjectIDDetector to MCRIDMapper
