+++ mycore-base/src/main/resources/config/mycore.properties
@@ -512,7 +512,7 @@ MCR.URIResolver.xslIncludes.functions=functions/acl.xsl,functions/classification
 ##############################################################################
 # implementation of the object id detector 
 # used for example in rest api to retrieve an MCRObjectID for a given id string
-MCR.Object.IDDetector.Class=org.mycore.frontend.iddetector.MCRDefaultObjectIDDetector
+MCR.Object.IDMapper.Class=org.mycore.frontend.idmapper.MCRDefaultIDMapper
 
 ##############################################################################
 # Static content settings
