+++ mycore-restapi/src/main/resources/components/restapi/config/mycore.properties
@@ -23,10 +23,10 @@ MCR.RestAPI.V2.Resource.Packages=org.mycore.restapi.v2
 #properties deprecated in 2024.06 - code was moved to mycore-solr - use new properties
 MCR.RestAPI.V2.AlternativeIdentifier.Objects.Keys=
 MCR.RestAPI.V2.AlternativeIdentifier.Derivates.Keys=
-## MCR.RestAPI.V2.AlternativeIdentifier.Objects.Keys=doi,category,derivates,id
-## MCR.RestAPI.V2.AlternativeIdentifier.Derivates.Keys=category,derivateOrder
-#    MCR.Object.IDDetector.ObjectSolrFields=doi,category,derivates,id
-#    MCR.Object.IDDetector.DerivateSolrFields=category,derivateOrder
+# MCR.RestAPI.V2.AlternativeIdentifier.Objects.Keys=doi,category,derivates,id
+# MCR.RestAPI.V2.AlternativeIdentifier.Derivates.Keys=category,derivateOrder
+##MCR.Object.IDMapper.ObjectSolrFields=doi,category,derivates,id
+##MCR.Object.IDMapper.ObjectSolrFields=category,derivateOrder
 
 MCR.Jersey.Resource.ApplicationPaths=%MCR.Jersey.Resource.ApplicationPaths%,api
