+++ mycore-solr/src/main/resources/components/solr/config/mycore.properties
@@ -100,6 +100,6 @@ MCR.Solr.AddDerivates.Fields=id,returnId,derivateMaindoc,iviewFile,derivateType
 #MCR.Solr.Disallowed.Facets=
 
 # detector, that retrieves an MCRObjectID from string 
-MCR.Object.IDDetector.Class=org.mycore.solr.iddetector.MCRSolrObjectIDDetector
-#MCR.Object.IDDetector.ObjectSolrFields=doi,category,derivates,id
-#MCR.Object.IDDetector.DerivateSolrFields=category,derivateOrder
+MCR.Object.IDMapper.Class=org.mycore.solr.idmapper.MCRSolrIDMapper
+#MCR.Object.IDMapper.ObjectSolrFields=doi,category,derivates,id
+#MCR.Object.IDMapper.DerivateSolrFields=category,derivateOrder
